pyproject.toml
[tool.poetry]
name = "valheim-server"
version = "0.1.0"
description = ""
authors = ["Dev Chakraborty <devchakraborty@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
aiozipstream = "^0.4"
aiofiles = "^0.6.0"
aiopath = "0.5.x"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
